feat(ph-cmd): add new upgrade command
