+++ clis/ph-cmd/package.json
@@ -38,7 +38,10 @@
     "change-case": "^5.4.4",
     "cmd-ts": "^0.14.3",
     "commander": "^12.1.0",
-    "enquirer": "^2.4.1"
+    "detect-package-manager": "^3.0.2",
+    "enquirer": "^2.4.1",
+    "read-pkg": "^9.0.1",
+    "write-package": "^7.2.0"
   },
   "devDependencies": {
     "vitest": "^3.2.4"
