+++ packages/codegen/src/file-builders/boilerplate/constants.ts
@@ -1,27 +1,3 @@
-export const POWERHOUSE_ORG = "@powerhousedao";
-
-/** Special packages that don't use the @powerhousedao organization */
-export const SPECIAL_PACKAGES = [
-  "document-model",
-  "document-drive",
-  "@renown/sdk",
-];
-
-/** Packages to exclude from version resolution (external dependencies) */
-export const EXCLUDED_PACKAGES = [
-  "@powerhousedao/document-engineering",
-  "@powerhousedao/scalars",
-  "@powerhousedao/diff-analyzer",
-  "@powerhousedao/analytics-engine-core",
-  "@powerhousedao/analytics-engine-graphql",
-  "@powerhousedao/analytics-engine-pg",
-  "@powerhousedao/analytics-engine-browser",
-  "@powerhousedao/analytics-engine-knex",
-];
-
-/** Version tags that should be resolved to actual versions */
-export const VERSION_TAGS = ["dev", "staging", "latest"];
-
 export const VERSIONED_DEPENDENCIES = [
   "@powerhousedao/common",
   "@powerhousedao/design-system",
@@ -29,6 +5,7 @@ export const VERSIONED_DEPENDENCIES = [
   "@powerhousedao/builder-tools",
   "document-model",
 ];
+
 export const VERSIONED_DEV_DEPENDENCIES = [
   "@powerhousedao/codegen",
   "@powerhousedao/config",
